switch over to using a cache-based approach for the resolver + clean up layout + add builder struct
