+++ lib/saluki-context/Cargo.toml
@@ -7,9 +7,9 @@ repository = { workspace = true }
 
 [dependencies]
 ahash = { workspace = true }
-indexmap = { workspace = true, features = ["std"] }
 metrics = { workspace = true }
 quick_cache = { workspace = true, features = ["ahash", "parking_lot"] }
+saluki-error = { workspace = true }
 saluki-metrics = { workspace = true }
 stringtheory = { workspace = true }
